requirements.txt
streamlit
pandas
pillow
boto3
azure-storage-blob
python-dotenv